Convert CurveSpec.js -> Curve.spec.tsx
